+++ docs/_app/immutable/chunks/3-37762b27.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/demo/code/_page.svelte-9b737ec2.js";export{t as component};