+++ docs/_app/immutable/chunks/4-bb436ee3.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/demo/complex/_page.svelte-26c238ec.js";export{t as component};