+++ docs/_app/immutable/chunks/5-a981e819.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/demo/iiif/_page.svelte-440eb703.js";export{t as component};