+++ docs/_app/immutable/chunks/7-32f20c41.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/scrollership/_page.svelte-8b309da4.js";export{t as component};