+++ src/lib/Meta.svelte
@@ -11,7 +11,6 @@
     </title>
   {/if}
   {#each Object.keys(data) as key}
-    <h1>{key}</h1>
-    <Elements elems={data[key]['c']} {settings}></Elements>
+      
   {/each}
 </svelte:head>